requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib[bcrypt]
psycopg2-binary
pydantic[email]
python-jose
python-multipart
alembic
boto3
pydantic[dotenv]
pytest
requests
celery
flower
redis
fastapi-mail
asyncio
pyqrcode
pypng
fastapi-utils